pyproject.toml
[project]
name = "arquiusbot-2"
version = "1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "discord-py[speed]==2.5.2",
]